* commit 'e7c9a387fa190b786b1f7becfe99cb22a6f2934a':
  Fix 64bit build

+++ linker/linker.cpp
@@ -1877,7 +1877,7 @@ bool soinfo::PrelinkImage() {
         break;
       case DT_SYMENT:
         if (d->d_un.d_val != sizeof(ElfW(Sym))) {
-          DL_ERR("invalid DT_SYMENT: %d", d->d_un.d_val);
+          DL_ERR("invalid DT_SYMENT: %zd", static_cast<size_t>(d->d_un.d_val));
           return false;
         }
         break;
@@ -1932,7 +1932,7 @@ bool soinfo::PrelinkImage() {
         break;
       case DT_RELAENT:
         if (d->d_un.d_val != sizeof(ElfW(Rela))) {
-          DL_ERR("invalid DT_RELAENT: %d", d->d_un.d_val);
+          DL_ERR("invalid DT_RELAENT: %zd", static_cast<size_t>(d->d_un.d_val));
           return false;
         }
         break;
@@ -1954,7 +1954,7 @@ bool soinfo::PrelinkImage() {
         break;
       case DT_RELENT:
         if (d->d_un.d_val != sizeof(ElfW(Rel))) {
-          DL_ERR("invalid DT_RELENT: %d", d->d_un.d_val);
+          DL_ERR("invalid DT_RELENT: %zd", static_cast<size_t>(d->d_un.d_val));
           return false;
         }
         break;
